requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
pandas==2.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
psycopg2-binary==2.9.7
gunicorn==21.2.0
Pillow==10.0.1
tqdm